requirements.txt
Flask==3.0.0
gunicorn==22.0.0
pg8000==1.31.2
python-dotenv==1.0.0
Werkzeug==3.0.1
qrcode==7.4.2
